+++ src/containerapp/HISTORY.rst
@@ -4,6 +4,8 @@ Release History
 ===============
 upcoming
 ++++++
+* 'az containerapp function invocations': Update application insights query 
+* 'az containerapp function keys': Update minimum replica check 
 
 1.3.0b1
 ++++++